REFACTOR: リファクタリングの完了
