+++ src/scanner.rs
@@ -1,11 +1,11 @@
-use crate::token::{match_keywords, Token, TokenType};
+use crate::token::{match_keywords, Literal, Token, TokenType};
 
 /// `Scanner`は、入力された文字列をトークンの配列に解析するための構造体
-pub struct Scanner {
+struct Scanner {
     /// 入力文字列を保持する
     /// マルチバイトのUTF-8文字も安全に取り扱えるように char 型として保持する
     pub source: Vec<char>,
-    /// 入力文字列を該当するトークンに変換した配列を保持する
+    /// 字句解析した結果のトークンを保持する
     pub tokens: Vec<Token>,
     /// スキャン中のトークンの最初の文字の位置を指す
     pub start: usize,
@@ -15,16 +15,16 @@ pub struct Scanner {
     pub line: usize,
 }
 
-impl Default for Scanner {
-    fn default() -> Self {
-        Self::new()
-    }
+pub fn scan_tokens(input: &str) -> Result<Vec<Token>, String> {
+    let mut scanner = Scanner::new(input);
+    scanner.scan_tokens()?;
+    Ok(scanner.tokens)
 }
 
 impl Scanner {
-    pub fn new() -> Self {
+    fn new(input: &str) -> Self {
         Scanner {
-            source: vec![],
+            source: input.chars().collect(),
             tokens: vec![],
             start: 0,
             current: 0,
@@ -32,81 +32,89 @@ impl Scanner {
         }
     }
 
-    pub fn scan_tokens(&mut self, source: &str) {
-        self.source = source.chars().collect();
-
+    fn scan_tokens(&mut self) -> Result<(), String> {
         while !self.is_at_end() {
             self.start = self.current;
-            self.scan_token();
+            self.scan_token()?;
         }
 
-        self.add_token(TokenType::Eof, "".to_string());
+        self.tokens.push(Token {
+            ty: TokenType::Eof,
+            lexeme: vec![],
+            literal: None,
+            line: self.line,
+        });
+
+        Ok(())
     }
 
-    fn scan_token(&mut self) {
+    fn scan_token(&mut self) -> Result<(), String> {
         let c = self.advance();
         match c {
-            '(' => self.add_token(TokenType::LParan, "(".to_string()),
-            ')' => self.add_token(TokenType::RParan, ")".to_string()),
-            '{' => self.add_token(TokenType::LBrace, "{".to_string()),
-            '}' => self.add_token(TokenType::RBrace, "}".to_string()),
-            ',' => self.add_token(TokenType::Comma, ",".to_string()),
-            '.' => self.add_token(TokenType::Dot, ".".to_string()),
-            '-' => self.add_token(TokenType::Minus, "-".to_string()),
-            '+' => self.add_token(TokenType::Plus, "+".to_string()),
-            ';' => self.add_token(TokenType::SemiColon, ";".to_string()),
+            '{' => self.add_token(TokenType::LBrace),
+            '}' => self.add_token(TokenType::RBrace),
+            '(' => self.add_token(TokenType::LParan),
+            ')' => self.add_token(TokenType::RParan),
+            ',' => self.add_token(TokenType::Comma),
+            '.' => self.add_token(TokenType::Dot),
+            '-' => self.add_token(TokenType::Minus),
+            '+' => self.add_token(TokenType::Plus),
+            ';' => self.add_token(TokenType::SemiColon),
             '/' => {
                 if self.matches('/') {
                     while self.peek() != '\n' && !self.is_at_end() {
                         self.advance();
                     }
                 } else {
-                    self.add_token(TokenType::Slash, "/".to_string());
+                    self.add_token(TokenType::Slash)
                 }
             }
-            '*' => self.add_token(TokenType::Star, "*".to_string()),
+            '*' => self.add_token(TokenType::Star),
             '!' => {
                 if self.matches('=') {
-                    self.add_token(TokenType::BangEqual, "!=".to_string())
+                    self.add_token(TokenType::BangEqual)
                 } else {
-                    self.add_token(TokenType::Bang, "!".to_string())
-                };
+                    self.add_token(TokenType::Bang)
+                }
             }
             '=' => {
                 if self.matches('=') {
-                    self.add_token(TokenType::EqualEqual, "==".to_string())
+                    self.add_token(TokenType::EqualEqual)
                 } else {
-                    self.add_token(TokenType::Equal, "=".to_string())
-                };
-            }
-            '<' => {
-                if self.matches('=') {
-                    self.add_token(TokenType::LessEqual, "<=".to_string())
-                } else {
-                    self.add_token(TokenType::Less, "<".to_string())
-                };
+                    self.add_token(TokenType::Equal)
+                }
             }
             '>' => {
                 if self.matches('=') {
-                    self.add_token(TokenType::GreaterEqual, ">=".to_string())
+                    self.add_token(TokenType::GreaterEqual)
                 } else {
-                    self.add_token(TokenType::Greater, ">".to_string())
-                };
+                    self.add_token(TokenType::Greater)
+                }
             }
-            ' ' | '\r' | '\t' => (),
+            '<' => {
+                if self.matches('=') {
+                    self.add_token(TokenType::LessEqual)
+                } else {
+                    self.add_token(TokenType::Less)
+                }
+            }
+            ' ' | '\t' | '\r' => {}
             '\n' => {
                 self.line += 1;
             }
-            '"' => self.string(),
-            '0'..='9' => self.number(),
+            '"' => self.string()?,
             _ => {
-                if self.is_alpha(c) {
-                    self.identifier();
+                if is_digit(c) {
+                    self.number()?;
+                } else if is_alpha(c) {
+                    self.identifier()
                 } else {
-                    todo!()
+                    return Err(String::from(format!("invalid token: {c}")));
                 }
             }
         };
+
+        Ok(())
     }
 
     fn advance(&mut self) -> char {
@@ -115,15 +123,31 @@ impl Scanner {
         c
     }
 
-    fn add_token(&mut self, ty: TokenType, literal: String) {
+    /// ソースコードの終わりに達しているかどうかを判定します。
+    fn is_at_end(&self) -> bool {
+        self.current >= self.source.len()
+    }
+
+    fn add_token(&mut self, ty: TokenType) -> () {
         self.tokens.push(Token {
             ty,
-            literal,
-            lexeme: String::new(),
+            lexeme: self.source[self.start..self.current].to_vec(),
+            literal: None,
             line: self.line,
         })
     }
 
+    fn add_literal_token(&mut self, ty: TokenType, literal: Literal) -> () {
+        self.tokens.push(Token {
+            ty,
+            lexeme: self.source[self.start..self.current].to_vec(),
+            literal: Some(literal),
+            line: self.line,
+        })
+    }
+
+    /// 次の文字が期待したものであった場合に `true`` を返却し、文字を消費する
+    /// 期待したものではなかった場合は、文字を消費しない
     fn matches(&mut self, c: char) -> bool {
         if self.is_at_end() {
             return false;
@@ -139,17 +163,21 @@ impl Scanner {
 
     fn peek(&self) -> char {
         if self.is_at_end() {
+            '\0'
+        } else {
+            self.source[self.current]
+        }
+    }
+
+    fn peek_next(&self) -> char {
+        if self.current + 1 >= self.source.len() {
             return '\0';
         }
 
-        self.source[self.current]
+        self.source[self.current + 1]
     }
 
-    fn is_at_end(&self) -> bool {
-        self.current >= self.source.len()
-    }
-
-    fn string(&mut self) {
+    fn string(&mut self) -> Result<(), String> {
         while self.peek() != '"' && !self.is_at_end() {
             if self.peek() == '\n' {
                 self.line += 1;
@@ -158,243 +186,243 @@ impl Scanner {
         }
 
         if self.is_at_end() {
-            return;
+            return Err(String::from("Unterminated string"));
         }
 
         self.advance();
 
-        let value = self.source[self.start + 1..self.current - 1]
+        // "..." のうち最初と最後のダブルクォートを無視して、中身の文字列のみ抽出する
+        let literal = self.source[self.start + 1..self.current - 1]
             .iter()
-            .collect();
-        self.add_token(TokenType::String, value);
+            .collect::<String>();
+        self.add_literal_token(TokenType::String, Literal::Str(literal));
+
+        Ok(())
     }
 
-    fn number(&mut self) {
-        while self.is_digit(self.peek()) {
+    fn number(&mut self) -> Result<(), String> {
+        while is_digit(self.peek()) {
             self.advance();
         }
 
-        if self.peek() == '.' && self.is_digit(self.peek_next()) {
+        if self.peek() == '.' && is_digit(self.peek_next()) {
             self.advance();
-            while self.is_digit(self.peek()) {
+            while is_digit(self.peek()) {
                 self.advance();
             }
         }
 
-        let value = self.source[self.start..self.current].iter().collect();
-        self.add_token(TokenType::Number, value);
+        let value = self.source[self.start..self.current]
+            .iter()
+            .collect::<String>()
+            .parse()
+            .map_err(|err| format!("invalid number: {err}"))?;
+        self.add_literal_token(TokenType::Number, Literal::Number(value));
+
+        Ok(())
     }
 
     fn identifier(&mut self) {
-        while self.is_alpha_numeric(self.peek()) {
+        while is_alpha_numeric(self.peek()) {
             self.advance();
         }
 
-        let text = self.source[self.start..self.current]
+        let literal = self.source[self.start..self.current]
             .iter()
             .collect::<String>();
-        let ty = match_keywords(&text);
-        self.add_token(ty, text);
-    }
-
-    fn is_digit(&self, c: char) -> bool {
-        c.is_ascii_digit()
-    }
-
-    fn is_alpha(&self, c: char) -> bool {
-        c.is_ascii_alphabetic()
-    }
-
-    fn is_alpha_numeric(&self, c: char) -> bool {
-        self.is_alpha(c) || self.is_digit(c)
-    }
-
-    fn peek_next(&self) -> char {
-        if self.current + 1 >= self.source.len() {
-            return '\0';
+        match match_keywords(&literal) {
+            Some(ty) => self.add_token(ty),
+            None => self.add_literal_token(TokenType::Identifier, Literal::Identifier(literal)),
         }
-        self.source[self.current + 1]
     }
+}
+
+fn is_digit(c: char) -> bool {
+    c.is_ascii_digit()
+}
+
+fn is_alpha(c: char) -> bool {
+    c.is_ascii_alphabetic()
+}
+
+fn is_alpha_numeric(c: char) -> bool {
+    is_digit(c) || is_alpha(c)
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::token::{Token, TokenType};
-
-    use super::Scanner;
+    use crate::{
+        scanner::scan_tokens,
+        token::TokenType,
+        token::{Literal, Token},
+    };
 
     #[test]
     fn test_one_char_token() {
-        let input = "(){},.-+;/*";
+        let input = "{}(),.-+;/*";
 
         let expected = vec![
             Token {
-                ty: TokenType::LParan,
-                literal: "(".to_string(),
-                lexeme: String::new(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::RParan,
-                literal: ")".to_string(),
-                lexeme: String::new(),
-                line: 1,
-            },
-            Token {
                 ty: TokenType::LBrace,
-                literal: "{".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['{'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::RBrace,
-                literal: "}".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['}'],
+                literal: None,
+                line: 1,
+            },
+            Token {
+                ty: TokenType::LParan,
+                lexeme: vec!['('],
+                literal: None,
+                line: 1,
+            },
+            Token {
+                ty: TokenType::RParan,
+                lexeme: vec![')'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Comma,
-                literal: ",".to_string(),
-                lexeme: String::new(),
+                lexeme: vec![','],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Dot,
-                literal: ".".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['.'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Minus,
-                literal: "-".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['-'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Plus,
-                literal: "+".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['+'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::SemiColon,
-                literal: ";".to_string(),
-                lexeme: String::new(),
+                lexeme: vec![';'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Slash,
-                literal: "/".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['/'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Star,
-                literal: "*".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['*'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Eof,
-                literal: "".to_string(),
-                lexeme: String::new(),
+                lexeme: vec![],
+                literal: None,
                 line: 1,
             },
         ];
 
-        let mut scanner = Scanner::new();
-        scanner.scan_tokens(input);
+        let tokens = scan_tokens(input).expect("スキャンに失敗しました。");
+        assert_eq!(
+            expected.len(),
+            tokens.len(),
+            "トークンの数が期待と異なります。"
+        );
 
-        for (idx, token) in scanner.tokens.into_iter().enumerate() {
-            let exp_token = &expected[idx];
+        for (expected_token, actual_token) in expected.into_iter().zip(tokens.into_iter()) {
             assert_eq!(
-                token.ty, exp_token.ty,
-                "tokens[{idx}] ty - got={}, expected={}",
-                token.ty, exp_token.ty,
-            );
-            assert_eq!(
-                token.literal, exp_token.literal,
-                "tokens[{idx}] literal - got={}, expected={}",
-                token.literal, exp_token.literal
+                expected_token, actual_token,
+                "期待するトークンと実際のトークンが異なります。"
             );
         }
     }
 
     #[test]
     fn test_conditional_char_token() {
-        let input = "!!=<<=>>====";
+        let input = "!!====>>=<<=";
 
         let expected = vec![
             Token {
                 ty: TokenType::Bang,
-                literal: "!".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['!'],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::BangEqual,
-                literal: "!=".to_string(),
-                lexeme: String::new(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Less,
-                literal: "<".to_string(),
-                lexeme: String::new(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::LessEqual,
-                literal: "<=".to_string(),
-                lexeme: String::new(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Greater,
-                literal: ">".to_string(),
-                lexeme: String::new(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::GreaterEqual,
-                literal: ">=".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['!', '='],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::EqualEqual,
-                literal: "==".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['=', '='],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Equal,
-                literal: "=".to_string(),
-                lexeme: String::new(),
+                lexeme: vec!['='],
+                literal: None,
+                line: 1,
+            },
+            Token {
+                ty: TokenType::Greater,
+                lexeme: vec!['>'],
+                literal: None,
+                line: 1,
+            },
+            Token {
+                ty: TokenType::GreaterEqual,
+                lexeme: vec!['>', '='],
+                literal: None,
+                line: 1,
+            },
+            Token {
+                ty: TokenType::Less,
+                lexeme: vec!['<'],
+                literal: None,
+                line: 1,
+            },
+            Token {
+                ty: TokenType::LessEqual,
+                lexeme: vec!['<', '='],
+                literal: None,
                 line: 1,
             },
             Token {
                 ty: TokenType::Eof,
-                literal: "".to_string(),
-                lexeme: String::new(),
+                lexeme: vec![],
+                literal: None,
                 line: 1,
             },
         ];
 
-        let mut scanner = Scanner::new();
-        scanner.scan_tokens(input);
+        let tokens = scan_tokens(input).expect("スキャンに失敗しました。");
+        assert_eq!(
+            expected.len(),
+            tokens.len(),
+            "トークンの数が期待と異なります。"
+        );
 
-        for (idx, token) in scanner.tokens.into_iter().enumerate() {
-            let exp_token = &expected[idx];
+        for (expected_token, actual_token) in expected.into_iter().zip(tokens.into_iter()) {
             assert_eq!(
-                token.ty, exp_token.ty,
-                "tokens[{idx}] ty - got={}, expected={}",
-                token.ty, exp_token.ty,
-            );
-            assert_eq!(
-                token.literal, exp_token.literal,
-                "tokens[{idx}] literal - got={}, expected={}",
-                token.literal, exp_token.literal
+                expected_token, actual_token,
+                "期待するトークンと実際のトークンが異なります。"
             );
         }
     }
@@ -402,46 +430,43 @@ mod tests {
     #[test]
     fn test_comment_out() {
         let input = r#"
-        ( 
-            // コメントアウト
-        ) 
+        (
+          // コメントアウト
+        )
         "#;
 
         let expected = vec![
             Token {
                 ty: TokenType::LParan,
-                literal: "(".to_string(),
-                lexeme: String::new(),
-                line: 1,
+                lexeme: vec!['('],
+                literal: None,
+                line: 2,
             },
             Token {
                 ty: TokenType::RParan,
-                literal: ")".to_string(),
-                lexeme: String::new(),
-                line: 1,
+                lexeme: vec![')'],
+                literal: None,
+                line: 4,
             },
             Token {
                 ty: TokenType::Eof,
-                literal: "".to_string(),
-                lexeme: String::new(),
-                line: 1,
+                lexeme: vec![],
+                literal: None,
+                line: 5,
             },
         ];
 
-        let mut scanner = Scanner::new();
-        scanner.scan_tokens(input);
+        let tokens = scan_tokens(input).expect("スキャンに失敗しました。");
+        assert_eq!(
+            expected.len(),
+            tokens.len(),
+            "トークンの数が期待と異なります。"
+        );
 
-        for (idx, token) in scanner.tokens.into_iter().enumerate() {
-            let exp_token = &expected[idx];
+        for (expected_token, actual_token) in expected.into_iter().zip(tokens.into_iter()) {
             assert_eq!(
-                token.ty, exp_token.ty,
-                "tokens[{idx}] ty - got={}, expected={}",
-                token.ty, exp_token.ty,
-            );
-            assert_eq!(
-                token.literal, exp_token.literal,
-                "tokens[{idx}] literal - got={}, expected={}",
-                token.literal, exp_token.literal
+                expected_token, actual_token,
+                "期待するトークンと実際のトークンが異なります。"
             );
         }
     }
@@ -449,265 +474,133 @@ mod tests {
     #[test]
     fn test_string_literal() {
         let input = r#"
-        "string_value"
-
-        "string
-value"
+        "hello_world"
         "#;
 
         let expected = vec![
             Token {
                 ty: TokenType::String,
-                literal: "string_value".to_string(),
-                lexeme: String::new(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::String,
-                literal: "string\nvalue".to_string(),
-                lexeme: String::new(),
+                lexeme: vec![
+                    '"', 'h', 'e', 'l', 'l', 'o', '_', 'w', 'o', 'r', 'l', 'd', '"',
+                ],
+                literal: Some(Literal::Str("hello_world".to_string())),
                 line: 2,
             },
             Token {
                 ty: TokenType::Eof,
-                literal: "".to_string(),
-                lexeme: String::new(),
-                line: 1,
+                lexeme: vec![],
+                literal: None,
+                line: 3,
             },
         ];
 
-        let mut scanner = Scanner::new();
-        scanner.scan_tokens(input);
+        let tokens = scan_tokens(input).expect("スキャンに失敗しました。");
+        assert_eq!(
+            expected.len(),
+            tokens.len(),
+            "トークンの数が期待と異なります。"
+        );
 
-        for (idx, token) in scanner.tokens.into_iter().enumerate() {
-            let exp_token = &expected[idx];
+        for (expected_token, actual_token) in expected.into_iter().zip(tokens.into_iter()) {
             assert_eq!(
-                token.ty, exp_token.ty,
-                "tokens[{idx}] ty - got={}, expected={}",
-                token.ty, exp_token.ty,
-            );
-            assert_eq!(
-                token.literal, exp_token.literal,
-                "tokens[{idx}] literal - got={}, expected={}",
-                token.literal, exp_token.literal
+                expected_token, actual_token,
+                "期待するトークンと実際のトークンが異なります。"
             );
         }
     }
 
     #[test]
     fn test_number_literal() {
-        let input = "123;123.123";
+        let input = r#"
+        0.145
+        "#;
 
         let expected = vec![
             Token {
                 ty: TokenType::Number,
-                lexeme: String::new(),
-                literal: "123".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::SemiColon,
-                lexeme: String::new(),
-                literal: ";".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Number,
-                lexeme: String::new(),
-                literal: "123.123".to_string(),
-                line: 1,
+                lexeme: vec!['0', '.', '1', '4', '5'],
+                literal: Some(Literal::Number(0.145)),
+                line: 2,
             },
             Token {
                 ty: TokenType::Eof,
-                lexeme: String::new(),
-                literal: "".to_string(),
-                line: 1,
+                lexeme: vec![],
+                literal: None,
+                line: 3,
             },
         ];
 
-        let mut scanner = Scanner::new();
-        scanner.scan_tokens(input);
+        let tokens = scan_tokens(input).expect("スキャンに失敗しました。");
+        assert_eq!(
+            expected.len(),
+            tokens.len(),
+            "トークンの数が期待と異なります。"
+        );
 
-        for (idx, token) in scanner.tokens.into_iter().enumerate() {
-            let exp_token = &expected[idx];
+        for (expected_token, actual_token) in expected.into_iter().zip(tokens.into_iter()) {
             assert_eq!(
-                token.ty, exp_token.ty,
-                "tokens[{idx}] ty - got={}, expected={}",
-                token.ty, exp_token.ty,
-            );
-            assert_eq!(
-                token.literal, exp_token.literal,
-                "tokens[{idx}] literal - got={}, expected={}",
-                token.literal, exp_token.literal
+                expected_token, actual_token,
+                "期待するトークンと実際のトークンが異なります。"
             );
         }
     }
 
     #[test]
-    fn test_keywords() {
-        let input = r#"and;
+    fn test_keyword() {
+        let input = r#"
         var five = 5;
-
-        var fn = fun(x, y) {
-            return x + y
-        }        
         "#;
 
         let expected = vec![
             Token {
-                ty: TokenType::And,
-                lexeme: String::new(),
-                literal: "and".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::SemiColon,
-                lexeme: String::new(),
-                literal: ";".to_string(),
-                line: 1,
-            },
-            Token {
                 ty: TokenType::Var,
-                lexeme: String::new(),
-                literal: "var".to_string(),
-                line: 1,
+                lexeme: vec!['v', 'a', 'r'],
+                literal: None,
+                line: 2,
             },
             Token {
                 ty: TokenType::Identifier,
-                lexeme: String::new(),
-                literal: "five".to_string(),
-                line: 1,
+                lexeme: vec!['f', 'i', 'v', 'e'],
+                literal: Some(Literal::Identifier("five".to_string())),
+                line: 2,
             },
             Token {
                 ty: TokenType::Equal,
-                lexeme: String::new(),
-                literal: "=".to_string(),
-                line: 1,
+                lexeme: vec!['='],
+                literal: None,
+                line: 2,
             },
             Token {
                 ty: TokenType::Number,
-                lexeme: String::new(),
-                literal: "5".to_string(),
-                line: 1,
+                lexeme: vec!['5'],
+                literal: Some(Literal::Number(5.0)),
+                line: 2,
             },
             Token {
                 ty: TokenType::SemiColon,
-                lexeme: String::new(),
-                literal: ";".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Var,
-                lexeme: String::new(),
-                literal: "var".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Identifier,
-                lexeme: String::new(),
-                literal: "fn".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Equal,
-                lexeme: String::new(),
-                literal: "=".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Fun,
-                lexeme: String::new(),
-                literal: "fun".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::LParan,
-                lexeme: String::new(),
-                literal: "(".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Identifier,
-                lexeme: String::new(),
-                literal: "x".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Comma,
-                lexeme: String::new(),
-                literal: ",".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Identifier,
-                lexeme: String::new(),
-                literal: "y".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::RParan,
-                lexeme: String::new(),
-                literal: ")".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::LBrace,
-                lexeme: String::new(),
-                literal: "{".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Return,
-                lexeme: String::new(),
-                literal: "return".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Identifier,
-                lexeme: String::new(),
-                literal: "x".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Plus,
-                lexeme: String::new(),
-                literal: "+".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::Identifier,
-                lexeme: String::new(),
-                literal: "y".to_string(),
-                line: 1,
-            },
-            Token {
-                ty: TokenType::RBrace,
-                lexeme: String::new(),
-                literal: "}".to_string(),
-                line: 1,
+                lexeme: vec![';'],
+                literal: None,
+                line: 2,
             },
             Token {
                 ty: TokenType::Eof,
-                lexeme: String::new(),
-                literal: "".to_string(),
-                line: 1,
+                lexeme: vec![],
+                literal: None,
+                line: 3,
             },
         ];
 
-        let mut scanner = Scanner::new();
-        scanner.scan_tokens(input);
+        let tokens = scan_tokens(input).expect("スキャンに失敗しました。");
+        assert_eq!(
+            expected.len(),
+            tokens.len(),
+            "トークンの数が期待と異なります。"
+        );
 
-        for (idx, token) in scanner.tokens.into_iter().enumerate() {
-            let exp_token = &expected[idx];
+        for (expected_token, actual_token) in expected.into_iter().zip(tokens.into_iter()) {
             assert_eq!(
-                token.ty, exp_token.ty,
-                "tokens[{idx}] ty - got={}, expected={}",
-                token.ty, exp_token.ty,
-            );
-            assert_eq!(
-                token.literal, exp_token.literal,
-                "tokens[{idx}] literal - got={}, expected={}",
-                token.literal, exp_token.literal
+                expected_token, actual_token,
+                "期待するトークンと実際のトークンが異なります。"
             );
         }
     }
